Update subliminal to latest version (c3c0c45)
# Refactor some code
# Improves legendastv
# Fixes podnapisi url
