+++ lib/subliminal/__init__.py
@@ -9,12 +9,13 @@ __copyright__ = 'Copyright 2016, Antoine Bertin'
 import logging
 
 from .core import (AsyncProviderPool, ProviderPool, check_video, download_best_subtitles, download_subtitles,
-                   list_subtitles, provider_manager, refiner_manager, save_subtitles)
+                   list_subtitles, save_subtitles, scan_video, scan_videos)
 from .cache import region
 from .exceptions import Error, ProviderError
+from .extensions import provider_manager, refiner_manager
 from .providers import Provider
-from .score import compute_score
-from .subtitle import Subtitle
-from .video import SUBTITLE_EXTENSIONS, VIDEO_EXTENSIONS, Episode, Movie, Video, scan_video, scan_videos
+from .score import compute_score, get_scores
+from .subtitle import SUBTITLE_EXTENSIONS, Subtitle
+from .video import VIDEO_EXTENSIONS, Episode, Movie, Video
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
