+++ lib/subliminal/providers/addic7ed.py
@@ -11,7 +11,8 @@ from .. import __short_version__
 from ..cache import SHOW_EXPIRATION_TIME, region
 from ..exceptions import AuthenticationError, ConfigurationError, DownloadLimitExceeded, TooManyRequests
 from ..subtitle import Subtitle, fix_line_ending, guess_matches
-from ..video import Episode, sanitize
+from ..utils import sanitize
+from ..video import Episode
 
 logger = logging.getLogger(__name__)
