+++ lib/subliminal/providers/podnapisi.py
@@ -19,7 +19,8 @@ from . import Provider
 from .. import __short_version__
 from ..exceptions import ProviderError
 from ..subtitle import Subtitle, fix_line_ending, guess_matches
-from ..video import Episode, Movie, sanitize
+from ..utils import sanitize
+from ..video import Episode, Movie
 
 logger = logging.getLogger(__name__)
