+++ lib/subliminal/providers/tvsubtitles.py
@@ -13,7 +13,8 @@ from .. import __short_version__
 from ..cache import EPISODE_EXPIRATION_TIME, SHOW_EXPIRATION_TIME, region
 from ..exceptions import ProviderError
 from ..subtitle import Subtitle, fix_line_ending, guess_matches
-from ..video import Episode, sanitize
+from ..utils import sanitize
+from ..video import Episode
 
 logger = logging.getLogger(__name__)
