+++ lib/subliminal/utils.py
@@ -1,0 +1,122 @@
+# -*- coding: utf-8 -*-
+from datetime import datetime
+import hashlib
+import os
+import re
+import struct
+
+
+def hash_opensubtitles(video_path):
+    """Compute a hash using OpenSubtitles' algorithm.
+
+    :param str video_path: path of the video.
+    :return: the hash.
+    :rtype: str
+
+    """
+    bytesize = struct.calcsize(b'<q')
+    with open(video_path, 'rb') as f:
+        filesize = os.path.getsize(video_path)
+        filehash = filesize
+        if filesize < 65536 * 2:
+            return
+        for _ in range(65536 // bytesize):
+            filebuffer = f.read(bytesize)
+            (l_value,) = struct.unpack(b'<q', filebuffer)
+            filehash += l_value
+            filehash &= 0xFFFFFFFFFFFFFFFF  # to remain as 64bit number
+        f.seek(max(0, filesize - 65536), 0)
+        for _ in range(65536 // bytesize):
+            filebuffer = f.read(bytesize)
+            (l_value,) = struct.unpack(b'<q', filebuffer)
+            filehash += l_value
+            filehash &= 0xFFFFFFFFFFFFFFFF
+    returnedhash = '%016x' % filehash
+
+    return returnedhash
+
+
+def hash_thesubdb(video_path):
+    """Compute a hash using TheSubDB's algorithm.
+
+    :param str video_path: path of the video.
+    :return: the hash.
+    :rtype: str
+
+    """
+    readsize = 64 * 1024
+    if os.path.getsize(video_path) < readsize:
+        return
+    with open(video_path, 'rb') as f:
+        data = f.read(readsize)
+        f.seek(-readsize, os.SEEK_END)
+        data += f.read(readsize)
+
+    return hashlib.md5(data).hexdigest()
+
+
+def hash_napiprojekt(video_path):
+    """Compute a hash using NapiProjekt's algorithm.
+
+    :param str video_path: path of the video.
+    :return: the hash.
+    :rtype: str
+
+    """
+    readsize = 1024 * 1024 * 10
+    with open(video_path, 'rb') as f:
+        data = f.read(readsize)
+    return hashlib.md5(data).hexdigest()
+
+
+def sanitize(string):
+    """Sanitize a string to strip special characters.
+
+    :param str string: the string to sanitize.
+    :return: the sanitized string.
+    :rtype: str
+
+    """
+    # only deal with strings
+    if string is None:
+        return
+
+    # replace some characters with one space
+    string = re.sub('[-:\(\)\.]', ' ', string)
+
+    # remove some characters
+    string = re.sub('[\']', '', string)
+
+    # replace multiple spaces with one
+    string = re.sub('\s+', ' ', string)
+
+    # strip and lower case
+    return string.strip().lower()
+
+
+def sanitize_release_group(string):
+    """Sanitize a string that represents a `release_group` by stripping the square brackets information
+    :param str string: the release group to sanitize.
+    :return: the sanitized release group.
+    :rtype: str
+    """
+    # only deal with strings
+    if string is None:
+        return
+
+    # strip square brackets information
+    string = re.sub('\s*\[\w+\]', '', string)
+
+    # strip and lower case
+    return string.strip().lower()
+
+
+def timestamp(date):
+    """Get the timestamp of the `date`, python2/3 compatible
+
+    :param datetime.datetime date: the utc date.
+    :return: the timestamp of the date.
+    :rtype: float
+
+    """
+    return (date - datetime(1970, 1, 1)).total_seconds()
