+++ lib/subliminal/video.py
@@ -1,14 +1,9 @@
 # -*- coding: utf-8 -*-
 from __future__ import division
 from datetime import datetime, timedelta
-import hashlib
 import logging
 import os
-import re
-import struct
 
-from babelfish import Error as BabelfishError, Language
-from enzyme import MKV
 from guessit import guessit
 
 logger = logging.getLogger(__name__)
@@ -21,9 +16,6 @@ VIDEO_EXTENSIONS = ('.3g2', '.3gp', '.3gp2', '.3gpp', '.60d', '.ajp', '.asf', '.
                     '.mpe', '.mpeg', '.mpg', '.mpv', '.mpv2', '.mxf', '.nsv', '.nut', '.ogg', '.ogm', '.omf', '.ps',
                     '.qt', '.ram', '.rm', '.rmvb', '.swf', '.ts', '.vfw', '.vid', '.video', '.viv', '.vivo', '.vob',
                     '.vro', '.wm', '.wmv', '.wmx', '.wrap', '.wvx', '.wx', '.x264', '.xvid')
-
-#: Subtitle extensions
-SUBTITLE_EXTENSIONS = ('.srt', '.sub', '.smi', '.txt', '.ssa', '.ass', '.mpl')
 
 
 class Video(object):
@@ -77,12 +69,12 @@ class Video(object):
 
     @property
     def exists(self):
-        """Test whether the video exists."""
+        """Test whether the video exists"""
         return os.path.exists(self.name)
 
     @property
     def age(self):
-        """Age of the video."""
+        """Age of the video"""
         if self.exists:
             return datetime.utcnow() - datetime.utcfromtimestamp(os.path.getmtime(self.name))
 
@@ -227,331 +219,3 @@ class Movie(Video):
             return '<%s [%r]>' % (self.__class__.__name__, self.title)
 
         return '<%s [%r, %d]>' % (self.__class__.__name__, self.title, self.year)
-
-
-def search_external_subtitles(path, directory=None):
-    """Search for external subtitles from a video `path` and their associated language.
-
-    Unless `directory` is provided, search will be made in the same directory as the video file.
-
-    :param str path: path to the video.
-    :param str directory: directory to search for subtitles.
-    :return: found subtitles with their languages.
-    :rtype: dict
-
-    """
-    dirpath, filename = os.path.split(path)
-    dirpath = dirpath or '.'
-    fileroot, fileext = os.path.splitext(filename)
-    subtitles = {}
-    for p in os.listdir(directory or dirpath):
-        # keep only valid subtitle filenames
-        if not p.startswith(fileroot) or not p.endswith(SUBTITLE_EXTENSIONS):
-            continue
-
-        # extract the potential language code
-        language_code = p[len(fileroot):-len(os.path.splitext(p)[1])].replace(fileext, '').replace('_', '-')[1:]
-
-        # default language is undefined
-        language = Language('und')
-
-        # attempt to parse
-        if language_code:
-            try:
-                language = Language.fromietf(language_code)
-            except ValueError:
-                logger.error('Cannot parse language code %r', language_code)
-
-        subtitles[p] = language
-
-    logger.debug('Found subtitles %r', subtitles)
-
-    return subtitles
-
-
-def scan_video(path, subtitles=True, embedded_subtitles=True, subtitles_dir=None):
-    """Scan a video and its subtitle languages from a video `path`.
-
-    :param str path: existing path to the video.
-    :param bool subtitles: scan for subtitles with the same name.
-    :param bool embedded_subtitles: scan for embedded subtitles.
-    :param str subtitles_dir: directory to search for subtitles.
-    :return: the scanned video.
-    :rtype: :class:`Video`
-
-    """
-    # check for non-existing path
-    if not os.path.exists(path):
-        raise ValueError('Path does not exist')
-
-    # check video extension
-    if not path.endswith(VIDEO_EXTENSIONS):
-        raise ValueError('%s is not a valid video extension' % os.path.splitext(path)[1])
-
-    dirpath, filename = os.path.split(path)
-    logger.info('Scanning video %r in %r', filename, dirpath)
-
-    # guess
-    video = Video.fromguess(path, guessit(path))
-
-    # size and hashes
-    video.size = os.path.getsize(path)
-    if video.size > 10485760:
-        logger.debug('Size is %d', video.size)
-        video.hashes['opensubtitles'] = hash_opensubtitles(path)
-        video.hashes['thesubdb'] = hash_thesubdb(path)
-        video.hashes['napiprojekt'] = hash_napiprojekt(path)
-        logger.debug('Computed hashes %r', video.hashes)
-    else:
-        logger.warning('Size is lower than 10MB: hashes not computed')
-
-    # external subtitles
-    if subtitles:
-        video.subtitle_languages |= set(search_external_subtitles(path, directory=subtitles_dir).values())
-
-    # video metadata with enzyme
-    try:
-        if filename.endswith('.mkv'):
-            with open(path, 'rb') as f:
-                mkv = MKV(f)
-
-            # main video track
-            if mkv.video_tracks:
-                video_track = mkv.video_tracks[0]
-
-                # resolution
-                if video_track.height in (480, 720, 1080):
-                    if video_track.interlaced:
-                        video.resolution = '%di' % video_track.height
-                    else:
-                        video.resolution = '%dp' % video_track.height
-                    logger.debug('Found resolution %s with enzyme', video.resolution)
-
-                # video codec
-                if video_track.codec_id == 'V_MPEG4/ISO/AVC':
-                    video.video_codec = 'h264'
-                    logger.debug('Found video_codec %s with enzyme', video.video_codec)
-                elif video_track.codec_id == 'V_MPEG4/ISO/SP':
-                    video.video_codec = 'DivX'
-                    logger.debug('Found video_codec %s with enzyme', video.video_codec)
-                elif video_track.codec_id == 'V_MPEG4/ISO/ASP':
-                    video.video_codec = 'XviD'
-                    logger.debug('Found video_codec %s with enzyme', video.video_codec)
-            else:
-                logger.warning('MKV has no video track')
-
-            # main audio track
-            if mkv.audio_tracks:
-                audio_track = mkv.audio_tracks[0]
-                # audio codec
-                if audio_track.codec_id == 'A_AC3':
-                    video.audio_codec = 'AC3'
-                    logger.debug('Found audio_codec %s with enzyme', video.audio_codec)
-                elif audio_track.codec_id == 'A_DTS':
-                    video.audio_codec = 'DTS'
-                    logger.debug('Found audio_codec %s with enzyme', video.audio_codec)
-                elif audio_track.codec_id == 'A_AAC':
-                    video.audio_codec = 'AAC'
-                    logger.debug('Found audio_codec %s with enzyme', video.audio_codec)
-            else:
-                logger.warning('MKV has no audio track')
-
-            # subtitle tracks
-            if mkv.subtitle_tracks:
-                if embedded_subtitles:
-                    embedded_subtitle_languages = set()
-                    for st in mkv.subtitle_tracks:
-                        if st.language:
-                            try:
-                                embedded_subtitle_languages.add(Language.fromalpha3b(st.language))
-                            except BabelfishError:
-                                logger.error('Embedded subtitle track language %r is not a valid language', st.language)
-                                embedded_subtitle_languages.add(Language('und'))
-                        elif st.name:
-                            try:
-                                embedded_subtitle_languages.add(Language.fromname(st.name))
-                            except BabelfishError:
-                                logger.debug('Embedded subtitle track name %r is not a valid language', st.name)
-                                embedded_subtitle_languages.add(Language('und'))
-                        else:
-                            embedded_subtitle_languages.add(Language('und'))
-                    logger.debug('Found embedded subtitle %r with enzyme', embedded_subtitle_languages)
-                    video.subtitle_languages |= embedded_subtitle_languages
-            else:
-                logger.debug('MKV has no subtitle track')
-
-    except:
-        logger.exception('Parsing video metadata with enzyme failed')
-
-    return video
-
-
-def scan_videos(path, age=None, **kwargs):
-    """Scan `path` for videos and their subtitles.
-
-    :param str path: existing directory path to scan.
-    :param datetime.timedelta age: maximum age of the video.
-    :param \*\*kwargs: parameters for the :func:`scan_video` function.
-    :return: the scanned videos.
-    :rtype: list of :class:`Video`
-
-    """
-    # check for non-existing path
-    if not os.path.exists(path):
-        raise ValueError('Path does not exist')
-
-    # check for non-directory path
-    if not os.path.isdir(path):
-        raise ValueError('Path is not a directory')
-
-    # walk the path
-    videos = []
-    for dirpath, dirnames, filenames in os.walk(path):
-        logger.debug('Walking directory %s', dirpath)
-
-        # remove badly encoded and hidden dirnames
-        for dirname in list(dirnames):
-            if dirname.startswith('.'):
-                logger.debug('Skipping hidden dirname %r in %r', dirname, dirpath)
-                dirnames.remove(dirname)
-
-        # scan for videos
-        for filename in filenames:
-            # filter on videos
-            if not filename.endswith(VIDEO_EXTENSIONS):
-                continue
-
-            # skip hidden files
-            if filename.startswith('.'):
-                logger.debug('Skipping hidden filename %r in %r', filename, dirpath)
-                continue
-
-            # reconstruct the file path
-            filepath = os.path.join(dirpath, filename)
-
-            # skip links
-            if os.path.islink(filepath):
-                logger.debug('Skipping link %r in %r', filename, dirpath)
-                continue
-
-            # skip old files
-            if age and datetime.utcnow() - datetime.utcfromtimestamp(os.path.getmtime(filepath)) > age:
-                logger.debug('Skipping old file %r in %r', filename, dirpath)
-                continue
-
-            # scan video
-            try:
-                video = scan_video(filepath, **kwargs)
-            except ValueError:  # pragma: no cover
-                logger.exception('Error scanning video')
-                continue
-
-            videos.append(video)
-
-    return videos
-
-
-def hash_opensubtitles(video_path):
-    """Compute a hash using OpenSubtitles' algorithm.
-
-    :param str video_path: path of the video.
-    :return: the hash.
-    :rtype: str
-
-    """
-    bytesize = struct.calcsize(b'<q')
-    with open(video_path, 'rb') as f:
-        filesize = os.path.getsize(video_path)
-        filehash = filesize
-        if filesize < 65536 * 2:
-            return
-        for _ in range(65536 // bytesize):
-            filebuffer = f.read(bytesize)
-            (l_value,) = struct.unpack(b'<q', filebuffer)
-            filehash += l_value
-            filehash &= 0xFFFFFFFFFFFFFFFF  # to remain as 64bit number
-        f.seek(max(0, filesize - 65536), 0)
-        for _ in range(65536 // bytesize):
-            filebuffer = f.read(bytesize)
-            (l_value,) = struct.unpack(b'<q', filebuffer)
-            filehash += l_value
-            filehash &= 0xFFFFFFFFFFFFFFFF
-    returnedhash = '%016x' % filehash
-
-    return returnedhash
-
-
-def hash_thesubdb(video_path):
-    """Compute a hash using TheSubDB's algorithm.
-
-    :param str video_path: path of the video.
-    :return: the hash.
-    :rtype: str
-
-    """
-    readsize = 64 * 1024
-    if os.path.getsize(video_path) < readsize:
-        return
-    with open(video_path, 'rb') as f:
-        data = f.read(readsize)
-        f.seek(-readsize, os.SEEK_END)
-        data += f.read(readsize)
-
-    return hashlib.md5(data).hexdigest()
-
-
-def hash_napiprojekt(video_path):
-    """Compute a hash using NapiProjekt's algorithm.
-
-    :param str video_path: path of the video.
-    :return: the hash.
-    :rtype: str
-
-    """
-    readsize = 1024 * 1024 * 10
-    with open(video_path, 'rb') as f:
-        data = f.read(readsize)
-    return hashlib.md5(data).hexdigest()
-
-
-def sanitize_release_group(string):
-    """Sanitize a string that represents a `release_group` by stripping the square brackets information.
-
-    :param str string: the release group to sanitize.
-    :return: the sanitized release group.
-    :rtype: str
-    """
-    # only deal with strings
-    if string is None:
-        return
-
-    # strip square brackets information
-    string = re.sub('\s*\[\w+\]', '', string)
-
-    # strip and lower case
-    return string.strip().lower()
-
-
-def sanitize(string):
-    """Sanitize a string to strip special characters.
-
-    :param str string: the string to sanitize.
-    :return: the sanitized string.
-    :rtype: str
-
-    """
-    # only deal with strings
-    if string is None:
-        return
-
-    # replace some characters with one space
-    string = re.sub('[-:\(\)]', ' ', string)
-
-    # remove some characters
-    string = re.sub('[\'\.]', '', string)
-
-    # replace multiple spaces with one
-    string = re.sub('\s+', ' ', string)
-
-    # strip and lower case
-    return string.strip().lower()
